Added files via upload
I can't seem to get the timer to read.  The game model for the gameOverStatus is before the timer, and it won't read the variable.  Wondering if it will matter if we switch the order.  Tried to put in the splash: do we have to import anything for that? Threw in the popup for loser.

Little things to throw together tomorrow during class!